feat: Finish welcome email
